split density code in two files
One file with the functions in src/ and one script in script/
