+++ src/AURORA.jl
@@ -25,7 +25,8 @@ export rotating, load_scattering_matrices
 include("main.jl")
 export calculate_e_transport
 
-
+include("analysis.jl")
+export make_density_file
